+++ src/components/app/media/FileDetails/FileDetails-styles.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const FileDetails = styled('div')({});
+
+export const Image = styled('img')({
+  width: '100%',
+  height: 'auto',
+  '& + .Detail': {
+    marginTop: '40rem',
+  },
+});