+++ src/components/app/media/FileDetails/FileDetails-types.ts
@@ -1,0 +1,20 @@
+interface ImageFile {
+  id: string;
+  created: Date;
+  modified: Date;
+  type: string;
+  src: string;
+  size: number;
+  width: number;
+  height: number;
+  srcset: {
+    src: string;
+    size: number;
+    width: number;
+    height: number;
+  }[];
+}
+
+export interface ViewProps {
+  file: ImageFile;
+}